requirements.txt
adal==1.2.6
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
azure-common==1.1.26
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.1.0
azure-mgmt-storage==11.2.0
azureml-core==1.24.0.post2
backports.tempfile==1.0
backports.weakref==1.0.post1
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.5.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docker==4.4.4
drf-yasg==1.20.0
drf-yasg2==1.19.4
facebook-sdk==3.1.0
Faker==4.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
jeepney==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpickle==2.0.0
jsonschema==3.2.0
MarkupSafe==1.1.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.1.2
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-twitter==3.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
SecretStorage==3.3.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.4
websocket-client==0.58.0
